Speed up graph loading time

+++ Anything/Models/HistoricalRate.cs
@@ -9,6 +9,6 @@ namespace Anything.Models
     {
         public string Title { get; set; }
         public List<string> ShortDate { get; set; }
-        public List<float> Amount { get; set; }
+        public List<double> Amount { get; set; }
     }
 }
